requirements.txt
boto==2.43.0
bz2file==0.98
click==6.6
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.8
django-dotenv==1.4.1
djangorestframework==3.3.1
gensim==0.13.3
gunicorn==19.6.0
numpy==1.11.2
praw==4.0.0b21
prawcore==0.2.1
psycopg2==2.6.2
requests==2.11.1
schedule==0.3.2
six==1.10.0
smart-open==1.3.5
update-checker==0.12
whitenoise==3.2
